+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-1.1.12 changelogs:
-- Fix rare crash with Compose
+1.1.13 changelogs:
+- Significantly improve OneConfig performance
+- Add support for mouse binds
+- Add ukulib config support
+- Update Compose, Skiko